Separa repositórios do sgbd utilizado

+++ src/main/java/app/shop/adaptorsInterfaces/dto/ItemPedidoDto.java
@@ -1,6 +1,6 @@
 package app.shop.adaptorsInterfaces.dto;
 
 public class ItemPedidoDto {
-    public Integer codigo_produto;
+    public Long codigo_produto;
     public Integer quantidade;
 }
